requirements.txt
langchain==0.1.0
langchain_openai==0.0.2
langchain-community==0.0.11
chromadb==0.5.0
openai==1.7.0
python-dotenv 
tiktoken==0.5.2
pymongo==4.6.3
tqdm==4.66.1
gradio
fastapi
uvicorn
pandas
markdown
gunicorn
pypdf